requirements.txt
pySide2==5.15.2.1
pyyaml==6.0.1
psutil==5.9.5
setproctitle==1.3.2
rasa==3.6.5
pynput==1.7.6
pyspellchecker==0.7.2
safetensors==0.3.1
transformers==4.31.0
keyboard==0.13.5

winshell; sys_platform == 'win32'
pywin32; sys_platform == 'win32'
